[RAPTOR-8302,8246] split integration tests, update R env to 4.2.1 (#670)
