+++ docker/dropin_env_base_r/build_image.sh
@@ -9,7 +9,7 @@
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 IMAGE_NAME=datarobot/dropin-env-base-r
-IMAGE_TAG=ubuntu18.04-py3.7-jre11-drum1.8.0-mlops8.0.7
+IMAGE_TAG=ubuntu20.04-r4.2.1-py3.8-jre11-drum1.9.5-mlops8.1.3
 
 pwd
