+++ jenkins/test_inference_model_templates.groovy
@@ -1,5 +1,4 @@
 node('release-dev && memory-intense'){
-  checkout scm
   stage ('test_inference_model_templates') {
     checkout scm
 
@@ -8,23 +7,21 @@ node('release-dev && memory-intense'){
     }
 
     checkoutDataRobot()
-    withQuantum([
-        bash: '''\
-            set -exuo pipefail
-            pushd DataRobot
-            make update_env
-            export LRS_POLL_DELAY_SECONDS=1
-            export LRS_RETRIES=120
-            export CUSTOM_MODEL_PREDICT_MEM_LIMIT=2147483648
-            export CUSTOM_MODEL_PREDICT_MEM_REQUEST=2147483648
-            export EXECUTION_ENVIRONMENT_LIMIT=20
-            export RESTRICT_EXECUTION_ENVIRONMENT_CREATION=false
-            export CLIENT_VERSION=\$(git rev-parse --short HEAD)
-            ./start.sh --kubernetes-k3d --kubernetes-validate
-            popd
-        '''.stripIndent(),
-        venvName: "datarobot-dev"
-    ])
+    sh '''
+        set -exuo pipefail
+        pushd DataRobot
+        make update_env
+        export LRS_POLL_DELAY_SECONDS=1
+        export LRS_RETRIES=120
+        export CUSTOM_MODEL_PREDICT_MEM_LIMIT=2147483648
+        export CUSTOM_MODEL_PREDICT_MEM_REQUEST=2147483648
+        export EXECUTION_ENVIRONMENT_LIMIT=20
+        export RESTRICT_EXECUTION_ENVIRONMENT_CREATION=false
+        export CLIENT_VERSION=\$(git rev-parse --short HEAD)
+        ./start.sh --kubernetes-k3d --kubernetes-validate
+        popd
+    '''.stripIndent()
+
     withQuantum([
         bash: '''\
             set -exuo pipefail
