+++ public_dropin_environments/r_lang/Dockerfile
@@ -1,26 +1,17 @@
 # This is the default base image for use with user models and workflows.
 # It contains a variety of common useful data-science packages and tools.
-FROM datarobot/r-dropin-env-base
-
-RUN apt update -y && apt install -y software-properties-common && \
-    add-apt-repository -y ppa:deadsnakes/ppa && \
-    apt update -y && apt install -y python3.7 python3.7-dev
-
-RUN python3.7 -m pip install setuptools wheel virtualenv && \
-    python3.7 -m virtualenv /opt/v3.7
-
-ENV PATH="/opt/v3.7/bin:$PATH"
-
+FROM datarobot/dropin-env-base-r:ubuntu20.04-r4.2.1-py3.8-jre11-drum1.9.5-mlops8.1.3
 # Install the list of core requirements, e.g. numpy, pandas, flask, rpy2.
 # **Don't modify this file!**
 COPY dr_requirements.txt dr_requirements.txt
 
-# TODO: remove once pyarrow is pinned to the same version across the repo
-RUN python3 -m pip install pip==20.2.4
-# '--upgrade-strategy eager' will upgrade dependencies
-# according to package requirements or to the latest
+# '--upgrade-strategy eager' upgrades dependencies
+# according to the package requirements or to the latest
 RUN python3 -m pip install -U --upgrade-strategy eager -r dr_requirements.txt --no-cache-dir && \
     rm -rf dr_requirements.txt
+
+# required by rpy2 https://github.com/rpy2/rpy2/issues/874
+ENV LD_LIBRARY_PATH=/usr/local/lib/R/lib:/usr/local/lib:/usr/lib/x86_64-linux-gnu:/usr/lib/jvm/java-11-openjdk-amd64/lib/server
 
 ENV HOME=/opt CODE_DIR=/opt/code ADDRESS=0.0.0.0:8080
 WORKDIR ${CODE_DIR}
