+++ public_dropin_environments/r_lang/dr_requirements.txt
@@ -1,5 +1,5 @@
 numpy==1.19.5
 pandas==1.0.5
 rpy2<=3.3.6
-pyarrow==0.14.1
-datarobot-drum[R]==1.9.4
+pyarrow>=0.14.1,<=3.0.0
+datarobot-drum[R]==1.9.5
