+++ public_dropin_environments/r_lang/env_info.json
@@ -1,8 +1,8 @@
 {
   "id": "5ea850ca1d41c8173c2feef6",
-  "name": "[DataRobot] R 3.6.3 Drop-In",
+  "name": "[DataRobot] R 4.2.1 Drop-In",
   "description": "This template environment can be used to create artifact-only R custom models that use the caret library. Your custom model archive need only contain your model artifacts if you use the environment correctly.",
   "programmingLanguage": "r",
-  "environmentVersionId": "62c7120d575f78f613dbc2f8",
+  "environmentVersionId": "62cdca4f6820547739e245b7",
   "isPublic": true
 }
