+++ task_templates/1_transforms/4_r_transform_recipe/model-metadata.yaml
@@ -34,7 +34,7 @@ typeSchema:
     # specify if the task accepts missing values
     - field: contains_missing
       condition: EQUALS # only EQUALS is supported
-      value: SUPPORTED # can be one of FORBIDDEN, SUPPORTED
+      value: FORBIDDEN # can be one of FORBIDDEN, SUPPORTED
 
   output_requirements:
