+++ task_templates/2_estimators/8_r_anomaly_detection/README.md
@@ -1,0 +1,4 @@
+
+### To run locally using 'drum'
+Paths are relative to `datarobot-user-models` root:
+`drum fit --code-dir task_templates/2_estimators/8_r_anomaly_detection/ --input tests/testdata/juniors_3_year_stats_regression.csv --target-type anomaly`
