+++ task_templates/3_pipelines/3_r_lang/model-metadata.yaml
@@ -4,7 +4,7 @@
 name: r_lang
 type: training # training (for custom tasks) or inference (for custom inference models)
 environmentID: 5ea850ca1d41c8173c2feef6 # optional, only required when using "drum push" to select environment
-targetType: multiclass  # can be one of: transform, binary, regression, anomaly, multiclass, unstructured (unstructured is only available for inference models)
+targetType: regression  # can be one of: transform, binary, regression, anomaly, multiclass, unstructured (unstructured is only available for inference models)
 
 # This part below defines the task's schema. Using this schema, on the fly,
 # DataRobot will validate if the task requirements match neighboring tasks in a blueprint
@@ -34,4 +34,4 @@ typeSchema:
     # specify if the task accepts missing values
     - field: contains_missing
       condition: EQUALS # only EQUALS is supported
-      value: SUPPORTED # can be one of FORBIDDEN, SUPPORTED
+      value: FORBIDDEN # can be one of FORBIDDEN, SUPPORTED
