+++ tests/fixtures/fit_custom.R
@@ -53,7 +53,7 @@ fit <- function(X, y, output_dir, class_order=NULL, row_weights=NULL, ...){
     # Drop constant columns
     step_zv(all_predictors()) %>%
     # Numeric preprocessing
-    step_medianimpute(all_numeric()) %>%
+    step_impute_median(all_numeric()) %>%
     step_normalize(all_numeric(), -all_outcomes()) %>%
     # Categorical preprocessing
     step_other(all_nominal(), -all_outcomes()) %>%
